third_party/WebKit/LayoutTests/http/tests/inspector/cache-storage/cache-live-update-cache-content-expected.txt

Tests that cache content is marked dirty if it is modified.

Dumping CacheStorage tree:
    (empty)
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        (cache empty)
Added entry
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        1
Deleted entry
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        (cache empty)
